+++ Runtime/Events.cs
@@ -6,31 +6,48 @@ using Unity.Services.Analytics.Platform;
 using Unity.Services.Authentication.Internal;
 using Unity.Services.Core.Device.Internal;
 using UnityEngine;
-using UnityEngine.Networking;
 
 namespace Unity.Services.Analytics
 {
     public static partial class Events
     {
         const string s_CollectUrlPattern = "https://collect.analytics.unity3d.com/collect/api/project/{0}/{1}";
+        const string k_ConsentStatusPrefKey = "unity.services.analytics.consent_status";
+        const string k_ForgetCallingId = "com.unity.services.analytics.Events." + nameof(OptOut);
 
-        internal static IPlayerId PlayerId { get; set; }
-        internal static IInstallationId InstallId { get; set; }
+        private enum ConsentStatus
+        {
+            Unknown = 0,
+            Forgetting = 1,
+            OptedOut = 2
+        }
 
-        internal static Analytics.Internal.Buffer dataBuffer = new Analytics.Internal.Buffer();
+        internal static IPlayerId PlayerId { get; private set; }
+        internal static IInstallationId InstallId { get; private set; }
+
+        internal static IBuffer dataBuffer = new Internal.Buffer();
         internal static Dispatcher dataDispatcher = new Dispatcher(dataBuffer);
         
-        public static Analytics.Internal.Buffer Buffer
+        public static IBuffer Buffer
         {
             get { return dataBuffer; }
         }
+
+        /// <summary>
+        /// This is the URL for the Unity Analytics privacy policy. This policy page should
+        /// be presented to the user in a platform-appropriate way along with the ability to
+        /// opt out of data collection.
+        /// </summary>
+        public static readonly string PrivacyUrl = "https://unity3d.com/legal/privacy-policy";
         
-        static UnityWebRequestAsyncOperation s_Request;
         static string s_CollectURL;
         static string s_SessionID;
         static Data.StdCommonParams s_CommonParams = new Data.StdCommonParams();
         static Ua2CoreInitializeCallback s_coreGlue = new Ua2CoreInitializeCallback();
         static string s_StartUpCallingId = "com.unity.services.analytics.Events.Startup";
+
+        static AnalyticsForgetter s_AnalyticsForgetter;
+        static ConsentStatus s_ConsentStatus;
 
         static Events()
         {
@@ -42,21 +59,98 @@ namespace Unity.Services.Analytics
                 return;
             }
 
+            s_ConsentStatus = (ConsentStatus)PlayerPrefs.GetInt(k_ConsentStatusPrefKey);
+
             s_SessionID = Guid.NewGuid().ToString();
             
             s_CommonParams.ClientVersion = Application.version;
+            s_CommonParams.ProjectID = Application.cloudProjectId;
             s_CommonParams.GameBundleID = Application.identifier;
             s_CommonParams.Platform = Platform.Runtime.Name();
             s_CommonParams.BuildGuuid = Application.buildGUID;
             s_CommonParams.Idfv = DeviceIdentifiersInternal.Idfv;
-            
+
             SetVariableCommonParams();
             
             DeviceIdentifiersInternal.SetupIdentifiers();
+        }
+
+        internal static void Initialize(IInstallationId installId, IPlayerId playerId, string environment)
+        {
+            InstallId = installId;
+            PlayerId = playerId;
             
+            s_CollectURL = String.Format(s_CollectUrlPattern, Application.cloudProjectId, environment.ToLowerInvariant());
+
             #if UNITY_ANALYTICS_DEVELOPMENT
-            Debug.LogFormat("UA2 Setup\nSessionID:{0}", s_SessionID);
+            Debug.LogFormat("UA2 Setup\nCollectURL:{0}\nSessionID:{1}", s_CollectURL, s_SessionID);
             #endif
+
+            if (s_ConsentStatus != ConsentStatus.Unknown)
+            {
+                OptOut();
+            }
+        }
+
+        /// <summary>
+        /// Opts the user out of sending analytics.
+        /// All existing cached events and any subsequent events will be discarded immediately.
+        /// A final 'forget me' signal will be uploaded which will trigger purge of analytics data for this user from the back-end.
+        /// If this 'forget me' event cannot be uploaded immediately (e.g. due to network outage), it will be reattempted regularly
+        /// until successful upload is confirmed.
+        /// Consent status is stored in PlayerPrefs so that the opted-out status is maintained over app restart.
+        /// This action cannot be undone.
+        /// </summary>
+        public static void OptOut()
+        {
+            if (s_ConsentStatus == ConsentStatus.OptedOut)
+            {
+#if UNITY_ANALYTICS_EVENT_LOGS
+                Debug.Log("This user has opted out. Any cached events have been discarded and no more will be collected.");
+#endif
+                // We have already been forgotten and so do not need to send the ForgetMe signal
+                dataBuffer.ClearDiskCache();
+                dataBuffer = new BufferRevoked();
+                dataDispatcher = new Dispatcher(dataBuffer);
+                ContainerObject.DestroyContainer();
+            }
+            else
+            {
+#if UNITY_ANALYTICS_EVENT_LOGS
+                Debug.Log("This user has opted out and is in the process of being forgotten...");
+#endif
+                // We have revoked consent but have not yet sent the ForgetMe signal
+                // Thus we need to keep some of the dispatcher alive until that is done
+                s_ConsentStatus = ConsentStatus.Forgetting;
+                PlayerPrefs.SetInt(k_ConsentStatusPrefKey, (int)s_ConsentStatus);
+                PlayerPrefs.Save();
+
+                // Clear everything out of the real buffer and replace it with a dummy
+                // that will swallow all events and do nothing
+                dataBuffer.ClearBuffer();
+                dataBuffer = new BufferRevoked();
+                dataDispatcher = new Dispatcher(dataBuffer);
+
+                s_AnalyticsForgetter = new AnalyticsForgetter(s_CollectURL,
+                                                              InstallId.GetOrCreateIdentifier(),
+                                                              Internal.Buffer.SaveDateTime(DateTime.UtcNow),
+                                                              k_ForgetCallingId,
+                                                              ForgetMeEventUploaded);
+                s_AnalyticsForgetter.AttemptToForget();
+            }
+        }
+
+        static void ForgetMeEventUploaded()
+        {
+            ContainerObject.DestroyContainer();
+
+            s_ConsentStatus = ConsentStatus.OptedOut;
+            PlayerPrefs.SetInt(k_ConsentStatusPrefKey, (int)s_ConsentStatus);
+            PlayerPrefs.Save();
+
+#if UNITY_ANALYTICS_EVENT_LOGS
+            Debug.Log("User opted out successfully and has been forgotten!");
+#endif
         }
 
         /// <summary>
@@ -75,16 +169,7 @@ namespace Unity.Services.Analytics
             bool isTiny = false;
             #endif
 
-            Data.Generator.GameStarted(ref dataBuffer, DateTime.UtcNow, s_CommonParams, s_StartUpCallingId, "PlayerSettings.productGUID.ToString()", SystemInfo.operatingSystem, isTiny, Platform.DebugDevice.IsDebugDevice(), Locale.CurrentCulture().Name);
-        }
-
-        internal static void SetEnvironment(string environment)
-        {
-            s_CollectURL = String.Format(s_CollectUrlPattern, Application.cloudProjectId, environment.ToLowerInvariant());
-                
-            #if UNITY_ANALYTICS_DEVELOPMENT
-            Debug.LogFormat("UA2 Setup\nCollectURL:{0}", s_CollectURL);
-            #endif
+            Data.Generator.GameStarted(ref dataBuffer, DateTime.UtcNow, s_CommonParams, s_StartUpCallingId, Application.buildGUID, SystemInfo.operatingSystem, isTiny, Platform.DebugDevice.IsDebugDevice(), Locale.AnalyticsRegionLanguageCode());
         }
 
         internal static void NewPlayerEvent()
@@ -127,15 +212,22 @@ namespace Unity.Services.Analytics
                 
                 return;
             }
-            dataBuffer.UserID = InstallId.GetOrCreateIdentifier();
-            dataBuffer.SessionID = s_SessionID;
-            dataDispatcher.CollectUrl = s_CollectURL;
-            dataDispatcher.Flush();
+
+            if (s_ConsentStatus == ConsentStatus.Forgetting)
+            {
+                s_AnalyticsForgetter.AttemptToForget();
+            }
+            else if (s_ConsentStatus != ConsentStatus.OptedOut)
+            {
+                dataBuffer.UserID = InstallId.GetOrCreateIdentifier();
+                dataBuffer.SessionID = s_SessionID;
+                dataDispatcher.CollectUrl = s_CollectURL;
+                dataDispatcher.Flush();
+            }
         }
 
         static void SetVariableCommonParams()
         {
-            s_CommonParams.UserCountry = Analytics.Internal.Platform.UserCountry.Name();
             s_CommonParams.Idfa = DeviceIdentifiersInternal.Idfa;
             s_CommonParams.DeviceVolume = DeviceVolumeProvider.GetDeviceVolume();
             s_CommonParams.BatteryLoad = SystemInfo.batteryLevel;